+++ scraper/src/main/java/GroceryFamily/GroceryDad/Parser/BarboraParser.java
@@ -1,4 +1,4 @@
-package GroceryFamily.scraper.Parser;
+package GroceryFamily.GroceryDad.Parser;
 
 import GroceryFamily.GroceryElders.model.Measurement;
 import GroceryFamily.GroceryElders.model.Product;
