+++ scraper/src/main/java/GroceryFamily/GroceryDad/productCheckOnThePage/NameChecking.java
@@ -1,4 +1,4 @@
-package GroceryFamily.scraper.productCheckOnThePage;
+package GroceryFamily.GroceryDad.productCheckOnThePage;
 
 import GroceryFamily.GroceryElders.model.Product;
