+++ scraper/src/main/java/GroceryFamily/GroceryDad/scraper/PrismaScraper.java
@@ -1,5 +1,6 @@
-package GroceryFamily.scraper;
+package GroceryFamily.GroceryDad.scraper;
 
+import GroceryFamily.GroceryDad.GroceryDadConfig;
 import GroceryFamily.GroceryElders.domain.Currency;
 import GroceryFamily.GroceryElders.domain.Price;
 import GroceryFamily.GroceryElders.domain.PriceUnit;
@@ -29,7 +30,7 @@ class PrismaScraper extends Scraper {
     }
 
     @Override
-    protected void scrap(List<String> categories) {
+    void scrap(List<String> categories) {
         FileCache<Product> cache = cache(categories);
         open(config.uri);
         waitUntilPageLoads();
