+++ scraper/src/main/java/GroceryFamily/GroceryDad/service/WebScraperServiceDubl.java
@@ -1,4 +1,4 @@
-package GroceryFamily.scraper.service;
+package GroceryFamily.GroceryDad.service;
 //
 //
 //import com.example.FinalProject.Parser.BarboraParser;
