+++ scraper/src/main/java/GroceryFamily/GroceryDad/similarity/ProductFilter.java
@@ -1,4 +1,4 @@
-package GroceryFamily.scraper.similarity;
+package GroceryFamily.GroceryDad.similarity;
 
 import GroceryFamily.GroceryElders.model.Product;
