--- scraper/src/main/java/GroceryFamily/scraper/productCheckOnThePage/PriceChecking.java
@@ -1,4 +1,0 @@
-package GroceryFamily.scraper.productCheckOnThePage;
-
-public class PriceChecking {
-}
