+++ scraper/src/test/java/GroceryFamily/GroceryDad/GroceryDadTest.java
@@ -1,10 +1,10 @@
-package GroceryFamily.scraper;
+package GroceryFamily.GroceryDad;
 
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
-@Disabled
+@Disabled // todo: introduce non-interactive driver for tests
 @SpringBootTest(classes = GroceryDad.class)
 class GroceryDadTest {
     @Test
